requirements.txt
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
Flask==3.0.3
gunicorn==23.0.0
h11==0.14.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.5.3
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.1
psutil==5.7.2
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
win32-setctime==1.1.0
wsproto==1.2.0
webdriver-manager==4.0.2
pytest-shutil==1.7.0